refactor: 야구 게임 숫자와 컨트롤 숫자 validation 분리
